+++ src/config/index.js
@@ -1,11 +1,12 @@
 // import resolveAssetSource from 'react-native/Libraries/Image/resolveAssetSource'
 import defaultUrl from '@/resources/medias/Silence02s.mp3'
 import notificationIcon from '@/resources/images/notification.xhdpi.png'
+import notificationIcon2 from '@/resources/images/notification2.xhdpi.png'
 // const defaultUrl = resolveAssetSource(resourceDefaultUrl).uri
 
 export {
   defaultUrl,
   notificationIcon,
+  notificationIcon2,
 }
 // export const defaultUrl = require('@/resources/medias/Silence02s.mp3')
-